+++ OrdinaryCapsule/CHANGELOG.md
@@ -1,9 +1,16 @@
 # Ordinary Capsule Change Log
 
-## 1.0.1
+## 1.1.0 (August 24, 2022)
+
+### Added
+
+* Added config option to unlock recipe automatically.
+* Added `assets/items.json` to customize supported items and production times.
+
+### Changed
 
 * Reduced number of patches.
 
-## 1.0.0
+## 1.0.0 (August 21, 2022)
 
 * Initial Version